requirements.txt
psycopg2-binary
Werkzeug==2.2.2
Flask==2.1.0
gunicorn
waitress
python-geohash
Geohash
flask-cors
sqlalchemy<2.0
sqlalchemy-cockroachdb
flask-login
flask-wtf
flask-sqlalchemy

